+++ infrastructure_simulator/board.py
@@ -16,6 +16,10 @@ ITEM_CONFIG = [
         "items" : [ Ampel, TestBild ]
     }
 ]
+class NamedInstance:
+    def __init__(self,name,instance):
+        self.name = name
+        self.instance = instance
 
 class Board:
 
@@ -27,8 +31,8 @@ class Board:
 
         self.infrastructure = {}
         for item in config['items']:
-            print(item['id'],item['type'])
-            self.infrastructure[item['id']] = self._get_instance(item['type'])
+            print(item['id'],item['name'],item['type'])
+            self.infrastructure[item['id']] = NamedInstance(item['name'],self._get_instance(item['type']))
 
         
 
@@ -51,7 +55,8 @@ class Board:
 
             config['infrastructure'].append({
                 "id" : id,
-                "commands" : item.get_commands()
+                "name" : item.name,
+                "commands" : item.instance.get_commands()
             })
 
         return config
@@ -63,7 +68,7 @@ class Board:
         if id not in self.infrastructure:
             raise ValueError(f"Unknown Infrastructure {id}")
 
-        return self.infrastructure[id].generate_image()
+        return self.infrastructure[id].instance.generate_image()
 
 
     async def generate(self,refresh,id):
@@ -72,7 +77,7 @@ class Board:
             raise ValueError(f"Unknown Infrastructure {id}")
 
         while True:
-            svg_img = self.infrastructure[id].generate_image()
+            svg_img = self.infrastructure[id].instance.generate_image()
             yield (b'--frame\r\n' b'Content-Type: image/svg+xml\r\n\r\n' +
                 bytearray(svg_img,encoding="UTF-8") + b'\r\n')
             await asyncio.sleep(0.1)
@@ -82,4 +87,4 @@ class Board:
         if id not in self.infrastructure:
             raise ValueError(f"Unknown Infrastructure {id}")
 
-        return self.infrastructure[id].control(cmd)
+        return self.infrastructure[id].instance.control(cmd)